processing-service/requirements.txt
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
aio-pika==9.4.1
openai==1.75.0
pillow==10.3.0
httpx==0.27.0